Fix arguments with parenthesis causing an error on Windows
Because %* is expanded before anything, any arguments with parenthesis
will make the .cmd fail with the following message: \"" was unexpected
at this time

This might actually be a security vulnerability because with the rightly
crafted arguments, one would possibly execute arbirtrary commands, even
if the arguments were properly escaped.  Read
https://gist.github.com/satazor/4e21543e5cd380032c2b6b38c3700223 for a
more detailed explanation.

To solve this, %* was moved out of the condition.

PR-URL: https://github.com/npm/cmd-shim/pull/26
Credit: @satazor
Close: #26
Reviewed-by: @isaacs