Fix generated shell scripts for MSYS2/MINGW
It has been reported tons and tons of times that the generated shell
scripts simply fail in Git Bash of the Git for Windows project. The
reason is that the Cygwin-specific basedir workaround is also required
for MSYS2 (which is used internally by the Git Bash).

Let's be nice to everybody and include support for MSYS2.

As MSYS2 distinguishes between MSYS/MINGW modes, triggered by setting
the environment variable MSYSTEM to the values MSYS, MINGW32 and
MINGW64, and as those modes are reflected in the output of `uname -s`,
we need to look not only for MSYS but also for MINGW in the output of
`uname`.

This addresses https://github.com/yarnpkg/yarn/issues/5349 among other
issues.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>