Update all dependencies
The original intention was to only update minimap, as `npm audit`
complains about this dependency being vulnerable to a regexp attach, and
that warning got a little annoying.

The minimap dependency is an indirect one, via tap, so let's update tap.
This also gives us nice coloring of the test output!

While at it, simply update all dependencies, not just tap.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>