Work around quoted batch file names
Fix: https://github.com/npm/cmd-shim/issues/10

Read %~dp0 in a subroutine, so that it does not get an incorrect
value when the command name is quoted.

While we're at it, bring this module up to 100% test coverage,
and fix a few obvious bugs that full coverage uncovered.

BREAKING CHANGE: This requires an update to read-cmd-shim, because the
format of the command invocation changed, and it reads that out of the
generated batch script.